rename lid test files
