+++ tests/test_toolkitapi.cpp
@@ -10,6 +10,7 @@
 
 
 * TABLE OF CONTENTS AND STRUCTURE
+
   ->  Numerical Diff Functions
   ->  Non-Fixuture Unit Tests
   ->  Unit tests using the fixtures
